bounce size_t changes off linux
svn path=/trunk/yarp2/; revision=8920

+++ src/carriers/human_carrier/HumanStream.cpp
@@ -9,7 +9,7 @@
 
 #include "HumanStream.h"
 
-int HumanStream::read(const Bytes& b) {
+ssize_t HumanStream::read(const Bytes& b) {
     if (interrupting) { return -1; }
     while (inputCache.size() < (unsigned int)b.length()) {
         cout << "*** CHECK OTHER TERMINAL FOR SOMETHING TO TYPE:"
