+++ src/carriers/mjpeg_carrier/MjpegStream.h
@@ -104,7 +104,7 @@ public:
 
     virtual void write(const Bytes& b);
 
-    virtual int read(const Bytes& b);
+    virtual ssize_t read(const Bytes& b);
 
     virtual void interrupt() { 
         delegate->getInputStream().interrupt();
