+++ src/carriers/tcpros_carrier/TcpRosStream.h
@@ -103,7 +103,7 @@ public:
 
     virtual void write(const Bytes& b);
 
-    virtual int read(const Bytes& b);
+    virtual ssize_t read(const Bytes& b);
 
     virtual void interrupt() { 
         delegate->getInputStream().interrupt();
